requirements.txt
requests>=2.25.1
pandas>=1.3.0
matplotlib>=3.4.2
seaborn>=0.11.1
python-dateutil>=2.8.1
google-cloud-language>=2.4.0
numpy>=1.20.0 